Update the schema so that we can store discord channel tournament links
This change fixes a lot of bugs in the bot, re-writes the game logs method and makes it ~5x more efficient.

This change also now stores Server+Channel name on the discord tournament channel links for easy debugging on the other end. This should fix a lot of logging issues in discord channels.